Fix: Show previous submissions
This fix rewrites the custom-defined `Moment` function so it
behaves as expected.

The issue was caused by an issue in the custom-defined `Moment`
function in `view.js`. When the JS code wanted to create a
"card" for a challenge submission (under the "Awarded" category,
for example), calling this function would fail, the card wouldn't
be added, and no further categories (such as "Pending" and
"Incorrect" would be added.